+++ src/components/Text/TextTypes.tsx
@@ -1,0 +1,6 @@
+export enum TextDisplaySizes {
+  Default = "default",
+  Caption = "caption",
+  Tag = "tag",
+  Mini = "mini",
+}